// frontend/src/api/generated/models/costAllocationResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */

export interface CostAllocationResponse {
  id: string;
  period_id: string;
  period_name: string;
  category_id: string;
  category_name: string;
  amount: string;
  allocation_basis: string;
  allocations: unknown[];
  reason: string;
}